+++ packages/pigeon/platform_tests/test_plugin/ios/Classes/TestPlugin.swift
@@ -10,9 +10,15 @@ import UIKit
  * example/integration_test/.
  */
 public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
+  var flutterAPI: FlutterIntegrationCoreApi
+
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let plugin = TestPlugin()
+    let plugin = TestPlugin(binaryMessenger: registrar.messenger())
     HostIntegrationCoreApiSetup.setUp(binaryMessenger: registrar.messenger(), api: plugin)
+  }
+
+  init(binaryMessenger: FlutterBinaryMessenger) {
+    flutterAPI = FlutterIntegrationCoreApi(binaryMessenger: binaryMessenger)
   }
 
   // MARK: HostIntegrationCoreApi implementation
@@ -35,5 +41,38 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
 
   func createNestedString(string: String) -> AllTypesWrapper {
     return AllTypesWrapper(values: AllTypes(aString: string))
+  }
+
+  func sendMultipleTypes(aBool: Bool, anInt: Int32, aString: String) -> AllTypes {
+    let someThings = AllTypes(aBool: aBool, anInt: anInt, aString: aString)
+    return someThings
+  }
+
+  func echoInt(anInt: Int32) -> Int32 {
+    return anInt
+  }
+
+  func echoBool(aBool: Bool) -> Bool {
+    return aBool
+  }
+
+  func noopAsync(completion: @escaping () -> Void) {
+    completion()
+  }
+
+  func echoAsyncString(aString: String, completion: @escaping (String) -> Void) {
+    completion(aString)
+  }
+
+  func callFlutterNoop(completion: @escaping () -> Void) {
+    flutterAPI.noop() {
+      completion()
+    }
+  }
+
+  func callFlutterEchoString(aString: String, completion: @escaping (String) -> Void) {
+    flutterAPI.echoString(aString: aString) { flutterString in
+      completion(flutterString)
+    }
   }
 }