+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.2.10
+
+* Changes generated Java enum field to be final.
+
+## 4.2.9
+
+* [kotlin] Fixes a bug with some methods that return `void`.
+
 ## 4.2.8
 
 * Adds the ability to use `runWithOptions` entrypoint to allow external libraries to use the pigeon easier.